requirements.txt
cx-Oracle==7.1.0
goto-statement==1.2